feat(codebase-refactoring): [4.3] アダプターとヘルパー関数の分離
- adapters.go: inventoryProviderAdapterをroot_model.goから分離
- helpers.go: CreateStatsDataFromGameState等のヘルパー関数を分離
- root_model.go: 分離したコードを削除し、544行に削減

Phase 3 (root_model.goの分割) 完了
